--- src/pages/RegisterPage/index.tsx
@@ -1,13 +1,0 @@
-import React from 'react'
-
-import RegisterForm from '../../components/RegisterForm'
-
-const RegisterPage = () => {
-  return (
-    <>
-      <RegisterForm />
-    </>
-  )
-}
-
-export default RegisterPage
