+++ src/pages/signupPage/index.tsx
@@ -1,0 +1,13 @@
+import React from 'react'
+
+import SignupForm from '../../components/signupForm'
+
+const SignupPage = () => {
+  return (
+    <>
+      <SignupForm />
+    </>
+  )
+}
+
+export default SignupPage
